third_party/WebKit/LayoutTests/fast/forms/suggested-value.html

<!DOCTYPE html>
<html>
<body>
<p id="description">This test setting suggested values on an input element 
and a textarea element. The dump below should have the "suggested value"
instead of "initial value".</p>
<pre><input id="test" type="text" value="initial value"><textarea id="textarea"></textarea></pre>
<script src="../../resources/dump-as-markup.js"></script>
<script>

var input = document.getElementById('test');
var textarea = document.getElementById('textarea');
textarea.value = "initial value";
var result = document.getElementById('result');
if (!window.internals)
    testFailed('This test requires internals object');
else {
    input.focus();
    input.selectionStart = input.selectionEnd = 0;

    internals.setSuggestedValue(input, 'suggested value');
    internals.setSuggestedValue(textarea, 'suggested value');

    Markup.description(document.getElementById('description').textContent)

    function addTextResult(value) { input.parentNode.appendChild(document.createTextNode(value + ': ' + eval(value))); }
    addTextResult('input.value');
    addTextResult('internals.suggestedValue(input)');
    addTextResult('input.selectionStart');
    addTextResult('input.selectionEnd');
    addTextResult('textarea.value');
    addTextResult('internals.suggestedValue(textarea)');
    
    Markup.dump(input.parentNode);
}

</script>
</body>
</html>
