third_party/WebKit/LayoutTests/fast/forms/suggested-value-expected.txt

This test setting suggested values on an input element 
and a textarea element. The dump below should have the "suggested value"
instead of "initial value".
| <input>
|   id="test"
|   type="text"
|   value="initial value"
|   this.value="initial value"
|   <shadow:root>
|     <div>
|       id="inner-editor"
|       "suggested value"
| <textarea>
|   id="textarea"
|   this.value="initial value"
|   <shadow:root>
|     <div>
|       id="inner-editor"
|       "suggested value"
| "input.value: initial value"
| "internals.suggestedValue(input): suggested value"
| "input.selectionStart: 0"
| "input.selectionEnd: 0"
| "textarea.value: initial value"
| "internals.suggestedValue(textarea): suggested value"
